improve color system
separated light and dark theme in two files. Added theme creation guidelines.
