+++ src/screens/ActivityFormScreen/ActivityFormScreen.js
@@ -279,7 +279,7 @@ const ActivityFormScreen = withTheme(({ theme, route, navigation }) => {
   )
 
   return(
-    <View style={{flex: 1, backgroundColor: theme.colors.surface}}>
+    <View style={{flex: 1, backgroundColor: theme.colors.activityFormScreenBackground}}>
       <Header 
         title={activity?.name?activity.name:t('activityForm.headerTitle')} 
         left='back' navigation={navigation} 
@@ -288,7 +288,7 @@ const ActivityFormScreen = withTheme(({ theme, route, navigation }) => {
       <KeyboardAwareScrollView style={{flexGrow: 0}} overScrollMode='never' contentContainerStyle={{flexGrow: 1}} keyboardShouldPersistTaps='handled' >
         <TextInput 
           error={nameInputError} 
-          style={{paddingHorizontal: 15, paddingTop: 10, fontSize: 16, height: 55, backgroundColor: theme.colors.surface}} 
+          style={{paddingHorizontal: 15, paddingTop: 10, fontSize: 16, height: 55, backgroundColor: theme.colors.textInputBackground}} 
           mode= 'outlined' 
           label={t('activityForm.nameInputLabel')}
           value={name} 
@@ -304,7 +304,7 @@ const ActivityFormScreen = withTheme(({ theme, route, navigation }) => {
             </HelperText> : null
         }
         <TextInput 
-          style={{paddingHorizontal: 15, paddingTop: 10, paddingBottom: 10, fontSize: 16, backgroundColor: theme.colors.surface}} 
+          style={{paddingHorizontal: 15, paddingTop: 10, paddingBottom: 10, fontSize: 16, backgroundColor: theme.colors.textInputBackground}} 
           mode= 'outlined' 
           multiline={true}
           label={t('activityForm.descriptionInputLabel')}
