+++ src/screens/AddTasksScreen.js
@@ -31,7 +31,7 @@ const AddTasksScreen = ({ theme, navigation, addTodayTask, tasksAddedToday }) =>
   )
 
   return (
-    <View style={{flex: 1, backgroundColor: theme.colors.surface}}>
+    <View style={{flex: 1, backgroundColor: theme.colors.addTasksScreenBackground}}>
     <Header 
       title={t('addTasks.title')}
       left='back'
