+++ src/screens/ArchivedActivitiesScreen.js
@@ -73,7 +73,7 @@ const ArchivedActivitiesScreen = withTheme(({ theme, route, navigation }) => {
   )
 
   return (
-    <View style={{flex: 1, backgroundColor: theme.colors.surface}}>
+    <View style={{flex: 1, backgroundColor: theme.colors.archivedActivitiesScreenBackground}}>
       <Header title={t("archivedActivitiesScreen.title", {goalName: goal.name})} left='back' navigation={navigation} />
       <View style={{ flex: 1 }}>
         {thisGoalArchivedActivities.length > 0?
