+++ src/screens/ArchivedGoalsScreen.js
@@ -64,7 +64,7 @@ const ArchivedGoalsScreen = withTheme(({ theme, navigation }) => {
   )
       
   return(
-    <View style={{flex: 1, backgroundColor: theme.colors.surface}}>
+    <View style={{flex: 1, backgroundColor: theme.colors.archivedGoalsScreenBackground}}>
       <Header 
         title={t("archivedGoalsScreen.title")} left='back' navigation={navigation} 
       />
