+++ src/screens/GoalScreen.js
@@ -191,7 +191,7 @@ const GoalScreen = withTheme(({ activities, goal, navigation, theme }) => {
 
   return (
     <>
-      <View style={{flex: 1, backgroundColor: theme.colors.surface}}>
+      <View style={{flex: 1, backgroundColor: theme.colors.goalScreenBackground}}>
         <Header title={goal.name} left={headerIcon} navigation={navigation} buttons={headerButtons}/>
         {/* ArchivedWarning only shows if the goal is archived */}
         <ArchivedWarning goal={goal}/>
