+++ src/screens/GoalsScreen.js
@@ -125,7 +125,7 @@ const GoalsScreen = withTheme(({ theme, navigation, goals }) => {
   )
       
   return(
-    <View style={{flex: 1, backgroundColor: theme.colors.surface}}>
+    <View style={{flex: 1, backgroundColor: theme.colors.goalsScreenBackground}}>
       <Header 
         title={t('goals.headerTitle')} navigation={navigation} 
         buttons={
