+++ src/theme/light.js
@@ -1,0 +1,145 @@
+export const lightPlacements = {
+  // Header
+  headerContent: 'onPrimary',
+  headerBackground: 'primary',
+  statusBarBackground: 'primary30',
+
+  // Tab bar
+  tabBarActiveIcon: 'primary',
+  tabBarInactiveIcon: 'neutral80',
+  tabBarBackground: 'surface',
+
+  // TodayScreen
+  todayDueIcon: 'onBackground',
+  todayCompletedIcon: 'onBackground',
+
+  activityBackground: 'surface',
+  runningActivityBackground: 'accentContainer',
+
+  progressBarToday: 'accent',
+  progressBarWeek: 'accent30',
+  progressBarBackground: 'accent80',
+
+  completedWeekliesSelector: 'neutral60',
+
+  // SelectWeekliesScreen
+    // Checkboxes use values in TodayScreen
+  completedCheckbox: 'neutral60',
+  selectWeekliesActivityBackground: 'surface',
+  selectWeekliesChangedText: 'primary',
+  selectWeekliesSelectedActivityBackground: 'accentContainer',
+
+  // Stats
+  barChartSelectorSelectedBackground: 'accent50',
+  barChartSelectorSelectedText: 'onSurface',
+  barChartBar: 'accent',
+  heatmap1: 'accent80',
+  heatmap2: 'accent60',
+  heatmap3: 'accent50',
+  heatmap4: 'accent30',
+  heatmapSkipped: 'secondary95',
+  heatmapEmptyBackground: 'transparent',
+  heatmapEmptyBorder: 'secondary90',
+
+  // Calendar
+  weekDayLabel: 'neutral60',
+  weekBackground: 'secondary95',
+  dayProgressBar: 'accent70',
+  weekProgressBar: 'accent',
+  weekProgressBarBackground: 'secondary90',
+  weekDayNumber: 'onSurface',
+  weekPastDayNumber: 'neutral60',
+  calendarTodayHighlightBackground: 'primary60',
+  calendarTodayHighlightText: 'onAccent',
+  calendarSoftTodayHighlightText: 'primary30',
+  weekPressedDayBackGround: 'accent',
+  weekFailedDayBackGround: 'neutral80',
+  weekPressedBackground: 'accent90',
+  calendarLongPressBackground: 'accent60',
+
+  // Tutorial
+  todayItemHighlight: 'primaryContainer',
+  pulsatingHighlight: 'primary',
+  speechBubbleBackground: 'primaryContainer',
+
+  // Miscellaneous
+  helpIcon: 'primary',
+  activityIndicator: 'primary',
+  textInputBackground: 'surface',
+
+  // Info Card
+  infoCardContent: 'onPrimaryContainer',
+  infoCardBackground: 'primary95',
+
+  // TimeInput
+  timeInputRegular: 'onSurface',
+  timeInputSelected: 'primary',
+
+  // CalendarWeekViewScreen
+  weekViewGoalCircularProgress: 'primary',
+  weekViewGoalCircularProgressBackground: 'primaryContainer',
+
+  // SettingsScreen
+  settingValueText: 'accent30',
+
+  // TodayScreen
+  todayScreenBackground: 'background',
+
+  // ActivityDetailScreen
+  activityDetailTimeInputRunning: 'primary',
+
+  // WeekDaySelector
+  weekDaySelectorPressedBackground: 'accent',
+  weekDaySelectorPressedText: 'onAccent',
+
+  // value used by react-native-paper
+  text: 'onSurface',
+
+  // Screen Backgrounds
+  dialogBackground: 'surface',
+  addTasksScreenBackground: 'surface',
+  archivedActivitiesScreenBackground: 'surface',
+  archivedGoalsScreenBackground: 'surface',
+  calendarDayViewScreenBackground: 'surface',
+  calendarScreenBackground: 'surface',
+  goalWeekViewBackground: 'surface',
+  calendarWeekViewScreenBackground: 'surface',
+  goalFormScreenBackground: 'surface',
+  goalScreenBackground: 'surface',
+  goalsScreenBackground: 'surface',
+  selectWeeklyActivitiesScreenBackground: 'surface',
+  settingsScreenBackground: 'surface',
+  statsScreenBackground: 'surface',
+  activityDetailsScreenBackground: 'surface',
+  activityFormScreenBackground: 'surface',
+}
+
+export const lightAliases = {
+  primary: 'primary40',
+  onPrimary: 'primary100',
+  primaryContainer: 'primary90',
+  onPrimaryContainer: 'primary10',
+  secondary: 'secondary40',
+  onSecondary: 'secondary100',
+  secondaryContainer: 'secondary90',
+  onSecondaryContainer: 'secondary10',
+  accent: 'accent40',
+  onAccent: 'accent100',
+  accentContainer: 'accent90',
+  onAccentContainer: 'accent10',
+
+  background: 'secondary95',
+  onBackground: 'neutral10',
+
+  surface: 'secondary99',
+  onSurface: 'neutral10',
+  
+  surfaceVariant: 'neutral90',
+  onSurfaceVariant: 'neutral30',
+  outline: 'neutral50',
+
+  inverseSurface: 'neutral20',
+  inverseOnSurface: 'neutral95',
+  inversePrimary: 'primary80',
+  
+}
